requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
langchain==0.3.14
langchain-google-genai
langchain-groq==0.2.5
langchain-community==0.3.14
sentence-transformers
faiss-cpu
pymupdf==1.25.3
pandas==2.2.3
fpdf2==2.8.2
pydantic==2.10.4
